+++ app/src/main/java/edu/ucne/prioridadlh/MainActivity.kt
@@ -20,6 +20,8 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Edit
 import androidx.compose.material3.ElevatedCard
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -27,6 +29,7 @@ import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -41,14 +44,21 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
 import androidx.room.Room
 import edu.ucne.prioridadlh.data.local.database.PrioridadesDb
+import edu.ucne.prioridadlh.presentacion.navigation.Screen
+import edu.ucne.prioridadlh.presentacion.navigation.prioridadlhNavHost
 import edu.ucne.prioridadlh.ui.theme.PrioridadLHTheme
 import edu.ucne.prioridadlt.data.local.entities.PrioridadesEntity
 import kotlinx.coroutines.launch
 
 class MainActivity : ComponentActivity() {
     private  lateinit var prioridadesDb: PrioridadesDb
+    private  lateinit var prioridadList: List<PrioridadesEntity>
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -69,7 +79,10 @@ class MainActivity : ComponentActivity() {
                             .fillMaxSize()
                             .padding(innerPadding)
                     ){
-                        PrioridadScreen()
+                        val navController = rememberNavController()
+                        prioridadlhNavHost(
+                            navController,
+                            prioridadesDb = prioridadesDb)
 
                     }
                 }
@@ -77,214 +90,245 @@ class MainActivity : ComponentActivity() {
         }
     }
 
+//
+//    @Composable
+//    fun PrioridadScreen(
+//        goTicketList: () -> Unit
+//    ) {
+//        var descripcion by remember { mutableStateOf("") }
+//        var diasCompromiso by remember { mutableStateOf("") }
+//        var error by remember { mutableStateOf<String?>(null) }
+//
+//        val scope = rememberCoroutineScope()
+//
+//        Scaffold { innerPadding ->
+//            Column(
+//                modifier = Modifier
+//                    .fillMaxSize()
+//                    .padding(innerPadding)
+//                    .padding(16.dp)
+//
+//            ) {
+//                Text(
+//                    text = "Registro Prioridad",
+//                    style = MaterialTheme.typography.titleLarge,
+//                    modifier = Modifier
+//                        .padding(bottom = 16.dp)
+//                        .align(Alignment.CenterHorizontally)
+//                )
+//
+//                ElevatedCard(
+//                    modifier = Modifier
+//                        .fillMaxWidth()
+//                        .padding(vertical = 8.dp)
+//                ) {
+//                    Column(
+//                        modifier = Modifier
+//                            .fillMaxWidth()
+//                            .padding(16.dp)
+//                            .align(Alignment.CenterHorizontally)
+//
+//                    ) {
+//                        OutlinedTextField(
+//                            label = { Text("Descripción") },
+//                            value = descripcion,
+//                            onValueChange = { descripcion = it },
+//                            modifier = Modifier.fillMaxWidth(),
+//                            singleLine = true,
+//                            isError = error != null
+//                        )
+//                        OutlinedTextField(
+//                            label = { Text("Días Compromiso") },
+//                            value = diasCompromiso,
+//                            onValueChange = {
+//                                if (it.all { char -> char.isDigit() }) {
+//                                    diasCompromiso = it
+//                                }
+//                            },
+//                            modifier = Modifier.fillMaxWidth(),
+//                            singleLine = true,
+//                            isError = error != null,
+//                            keyboardOptions = KeyboardOptions.Default.copy(
+//                                keyboardType = KeyboardType.Number
+//                            )
+//                        )
+//                        if (error != null) {
+//                            Text(
+//                                text = error!!,
+//                                color = Color.Red,
+//                                style = MaterialTheme.typography.bodySmall,
+//                                modifier = Modifier.padding(vertical = 8.dp)
+//                            )
+//                        }
+//                        Spacer(modifier = Modifier.height(16.dp))
+//                        Row(
+//                            modifier = Modifier.fillMaxWidth(),
+//                            horizontalArrangement = Arrangement.SpaceEvenly
+//                        ) {
+//                            OutlinedButton(
+//                                onClick = {
+//                                    scope.launch {
+//                                        nuevasPrioridades()
+//                                    }
+//                                    goTicketList()
+//                                }
+//                            ) {
+//                                Icon(
+//                                    imageVector = Icons.Default.Add,
+//                                    contentDescription = "Nuevo"
+//                                )
+//                                Spacer(modifier = Modifier.width(0.4.dp))
+//                                Text(text = "Nuevo")
+//                            }
+//
+//                            OutlinedButton(
+//                                onClick = {
+//                                    when {
+//                                        descripcion.isBlank() -> error = "Coloque la Descripción"
+//                                        diasCompromiso.isBlank() -> error = "Coloque los Días Compromiso"
+//                                        diasCompromiso.toIntOrNull() == 0 -> error = "Los Días Compromiso no pueden ser 0"
+//                                        else -> {
+//                                            scope.launch {
+//                                                savePrioridad(
+//                                                    PrioridadesEntity(
+//                                                        Descripcion = descripcion,
+//                                                        DiasCompromiso = diasCompromiso
+//                                                    )
+//                                                )
+//                                                descripcion = ""
+//                                                diasCompromiso = ""
+//                                                error = null
+//                                            }
+//                                            goTicketList()
+//                                        }
+//                                    }
+//                                }
+//                            ) {
+//                                Icon(
+//                                    imageVector = Icons.Default.Edit,
+//                                    contentDescription = "Guardar"
+//                                )
+//                                Spacer(modifier = Modifier.width(0.4.dp))
+//                                Text(text = "Guardar")
+//                            }
+//                        }
+//                    }
+//                }
+//            }
+//        }
+//    }
+//
+//    @OptIn(ExperimentalMaterial3Api::class)
+//    @Composable
+//    fun PrioridadListSc(
+//        prioridadList: List<PrioridadesEntity>,
+//        onAddTicket: () -> Unit
+//    ) {
+//        Scaffold (
+//            modifier = Modifier.fillMaxSize(),
+//            topBar = {
+//                TopAppBar(
+//                    title = {
+//                        Row(
+//                        modifier = Modifier
+//                            .fillMaxWidth(),
+//                        verticalAlignment = Alignment.CenterVertically,
+//                        horizontalArrangement = Arrangement.Center
+//                        ) {
+//                            Text(text = "Lista de Prioridades")
+//                        }
+//                    }
+//
+//                )
+//            },
+//            floatingActionButton = {
+//                FloatingActionButton(onClick = onAddTicket) {
+//                    Icon(Icons.Filled.Add, "Agregar nueva entidad")
+//                }
+//            }){
+//            Column(
+//                modifier = Modifier.fillMaxSize()
+//                    .padding(it)
+//            ) {
+//                Spacer(modifier = Modifier.height(32.dp))
+//
+//                LazyColumn(
+//                    modifier = Modifier.fillMaxSize()
+//                ) {
+//                    items(prioridadList) { item ->
+//                        PrioridadRow(item)
+//                    }
+//                }
+//            }
+//
+//        }
+//
+//    }
+//
+//    @Composable
+//    private fun PrioridadRow(it: PrioridadesEntity){
+//        Row(
+//            verticalAlignment = Alignment.CenterVertically
+//        ){
+//            Text(modifier = Modifier.weight(0.2f), text = it.PrioridadId.toString())
+//            Text(
+//                modifier = Modifier.weight(0.5f),
+//                text = it.Descripcion,
+//            )
+//            Text(modifier = Modifier.weight(0.2f), text = it.DiasCompromiso)
+//        }
+//        HorizontalDivider()
+//    }
+//
+//
+//
+//    private suspend fun savePrioridad(prioridad: PrioridadesEntity){
+//        prioridadesDb.prioridadesDao().save(prioridad)
+//    }
+//
+//    private suspend fun nuevasPrioridades() {
+//        prioridadesDb.prioridadesDao().deleteAll()
+//    }
+//
+//
+//    @Composable
+//    fun prioridadlhNavHost(
+//        navHostController: NavHostController
+//    ){
+//        val lifecycleOwner = androidx.lifecycle.compose.LocalLifecycleOwner.current
+//        val prioridadList by prioridadesDb.prioridadesDao().getAll()
+//            .collectAsStateWithLifecycle(
+//                initialValue = emptyList(),
+//                lifecycleOwner = lifecycleOwner,
+//                minActiveState = Lifecycle.State.STARTED
+//            )
+//
+//        NavHost(
+//            navController = navHostController,
+//            startDestination = Screen.PrioridadesList)
+//        {
+//            composable<Screen.Prioridad> {
+//                PrioridadScreen(
+//                    goTicketList = { navHostController.navigate(Screen.PrioridadesList)
+//                    }
+//                )
+//
+//            }
+//
+//            composable<Screen.PrioridadesList> {
+//                PrioridadListSc(
+//                    prioridadList = prioridadList,
+//                    onAddTicket = { navHostController.navigate(Screen.Prioridad(0)) }
+//                )
+//            }
+//        }
+//    }
 
-    @Composable
-    fun PrioridadScreen() {
-        var descripcion by remember { mutableStateOf("") }
-        var diasCompromiso by remember { mutableStateOf("") }
-        var error by remember { mutableStateOf<String?>(null) }
-
-        val scope = rememberCoroutineScope()
-
-        Scaffold { innerPadding ->
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(innerPadding)
-                    .padding(16.dp)
-
-            ) {
-                Text(
-                    text = "Registro Prioridad",
-                    style = MaterialTheme.typography.titleLarge,
-                    modifier = Modifier
-                        .padding(bottom = 16.dp)
-                        .align(Alignment.CenterHorizontally)
-                )
-
-                ElevatedCard(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(vertical = 8.dp)
-                ) {
-                    Column(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(16.dp)
-                            .align(Alignment.CenterHorizontally)
-
-                    ) {
-                        OutlinedTextField(
-                            label = { Text("Descripción") },
-                            value = descripcion,
-                            onValueChange = { descripcion = it },
-                            modifier = Modifier.fillMaxWidth(),
-                            singleLine = true,
-                            isError = error != null
-                        )
-                        OutlinedTextField(
-                            label = { Text("Días Compromiso") },
-                            value = diasCompromiso,
-                            onValueChange = {
-                                if (it.all { char -> char.isDigit() }) {
-                                    diasCompromiso = it
-                                }
-                            },
-                            modifier = Modifier.fillMaxWidth(),
-                            singleLine = true,
-                            isError = error != null,
-                            keyboardOptions = KeyboardOptions.Default.copy(
-                                keyboardType = KeyboardType.Number
-                            )
-                        )
-                        if (error != null) {
-                            Text(
-                                text = error!!,
-                                color = Color.Red,
-                                style = MaterialTheme.typography.bodySmall,
-                                modifier = Modifier.padding(vertical = 8.dp)
-                            )
-                        }
-                        Spacer(modifier = Modifier.height(16.dp))
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.SpaceEvenly
-                        ) {
-                            OutlinedButton(
-                                onClick = {
-                                    scope.launch {
-                                        nuevasPrioridades()
-                                    }
-                                }
-                            ) {
-                                Icon(
-                                    imageVector = Icons.Default.Add,
-                                    contentDescription = "Nuevo"
-                                )
-                                Spacer(modifier = Modifier.width(0.4.dp))
-                                Text(text = "Nuevo")
-                            }
-
-                            OutlinedButton(
-                                onClick = {
-                                    when {
-                                        descripcion.isBlank() -> error = "Coloque la Descripción"
-                                        diasCompromiso.isBlank() -> error = "Coloque los Días Compromiso"
-                                        diasCompromiso.toIntOrNull() == 0 -> error = "Los Días Compromiso no pueden ser 0"
-                                        else -> {
-                                            scope.launch {
-                                                savePrioridad(
-                                                    PrioridadesEntity(
-                                                        Descripcion = descripcion,
-                                                        DiasCompromiso = diasCompromiso
-                                                    )
-                                                )
-                                                descripcion = ""
-                                                diasCompromiso = ""
-                                                error = null
-                                            }
-                                        }
-                                    }
-                                }
-                            ) {
-                                Icon(
-                                    imageVector = Icons.Default.Edit,
-                                    contentDescription = "Guardar"
-                                )
-                                Spacer(modifier = Modifier.width(0.4.dp))
-                                Text(text = "Guardar")
-                            }
-                        }
-                    }
-                }
-
-                val lifecycleOwner = androidx.lifecycle.compose.LocalLifecycleOwner.current
-                val prioridadList by prioridadesDb.prioridadesDao().getAll()
-                    .collectAsStateWithLifecycle(
-                        initialValue = emptyList(),
-                        lifecycleOwner = lifecycleOwner,
-                        minActiveState = Lifecycle.State.STARTED
-                    )
-                PrioridadListSc(prioridadList)
-            }
-        }
-    }
-
-    @Composable
-    fun PrioridadListSc(prioridadList: List<PrioridadesEntity>) {
-        Column(
-            modifier = Modifier.fillMaxSize()
-        ) {
-            Spacer(modifier = Modifier.height(16.dp))
-            Text(
-                text = "Listado",
-                style = MaterialTheme.typography.titleLarge,
-                modifier = Modifier.padding(start = 16.dp, bottom = 8.dp)
-            )
-
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 16.dp, vertical = 8.dp),
-                horizontalArrangement = Arrangement.Center
-            ) {
-                Text(
-                    text = "#",
-                    style = MaterialTheme.typography.titleSmall,
-                    modifier = Modifier.weight(0.2f)
-                )
-                Text(
-                    text = "Descripción",
-                    style = MaterialTheme.typography.titleSmall,
-                    modifier = Modifier.weight(0.5f)
-                )
-                Text(
-                    text = "Días",
-                    style = MaterialTheme.typography.titleSmall,
-                    modifier = Modifier.weight(0.2f)
-                )
-            }
-
-            LazyColumn(
-                modifier = Modifier.fillMaxSize()
-            ) {
-                items(prioridadList) { item ->
-                    PrioridadRow(item)
-                }
-            }
-        }
-    }
-
-    @Composable
-    private fun PrioridadRow(it: PrioridadesEntity){
-        Row(
-            verticalAlignment = Alignment.CenterVertically
-        ){
-            Text(modifier = Modifier.weight(0.2f), text = it.PrioridadId.toString())
-            Text(
-                modifier = Modifier.weight(0.5f),
-                text = it.Descripcion,
-            )
-            Text(modifier = Modifier.weight(0.2f), text = it.DiasCompromiso)
-        }
-        HorizontalDivider()
-    }
-
-
-
-    private suspend fun savePrioridad(prioridad: PrioridadesEntity){
-        prioridadesDb.prioridadesDao().save(prioridad)
-    }
-
-    private suspend fun nuevasPrioridades() {
-        prioridadesDb.prioridadesDao().deleteAll()
-    }
 
     @Preview(showBackground = true, showSystemUi = true)
     @Composable
     fun GreetingPreview() {
         PrioridadLHTheme {
-            PrioridadScreen()
         }
     }
 }
