+++ app/src/main/java/edu/ucne/prioridadlh/presentacion/navigation/Screen.kt
@@ -1,0 +1,10 @@
+package edu.ucne.prioridadlh.presentacion.navigation
+
+import kotlinx.serialization.Serializable
+
+sealed class Screen{
+    @Serializable
+    data object  PrioridadesList : Screen()
+    @Serializable
+    data class Prioridad(val PrioridadId: Int) : Screen()
+}
